# apps/gateway/src/main/java/com/coldguard/gateway/api/user/Role.java
package com.coldguard.gateway.api.user;

/** The platform roles, as REST clients name them. */
public enum Role {
  OPERATIONS_SUPERVISOR,
  OPERATOR,
  MAINTENANCE_TECHNICIAN,
  AUDITOR,
  PLATFORM_ADMIN
}