# contracts/rest/openapi.yaml
openapi: 3.0.3
info:
  title: ColdGuard API
  version: 0.3.0
  description: |
    REST API of the ColdGuard Gateway, the only entry point to the platform. It is the contract the
    frontend integrates against.

    **Every operation carries `x-status`:**

    - `implemented`: it exists in the Gateway today and a test keeps this file and the code in sync.
    - `planned`: it is specified here so the frontend can build against it (with a mock) before the
      backend exists. Its shape may still change; it moves to `implemented` when its spec lands.

    `x-roles` lists the roles allowed to call it (the source of truth is
    `docs/security/authn-authz.md`; a test checks that both agree with the real security chain).

    **Conventions**

    - Authentication: `POST /auth/login` returns a Bearer token valid for one hour; there is no
      refresh token, so a `401` means "sign in again".
    - JSON in camelCase, instants in ISO-8601 UTC, enums without any protocol prefix.
    - Lists: `?page=0&size=20` returns `{items, page}`; time-ordered ones use a cursor,
      `?cursor=…&size=…`, and return `{items, nextCursor, hasMore}`. `size=0` or omitted selects the
      server default; the maximum is 100.
    - Errors are Problem Details (RFC 9457) with a stable machine-readable `code` and a
      `correlationId`, also sent as the `X-Correlation-Id` response header. Send that header yourself
      to correlate a request.
...

...
      commands are never retried.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Create a user
      operationId: createUser
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get a user
      operationId: getUser
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{userId}/roles:
    post:
      tags: [Users]
      summary: Assign a role
      description: A reason is required. Assigning a role the user already has changes nothing.
      operationId: assignRole
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleChangeRequest' }
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      summary: Revoke a role
      description: |
        The role and the reason travel in the request body. The last enabled administrator cannot lose
        the `PLATFORM_ADMIN` role (`409`, `USER_STATE_CONFLICT`).
      operationId: revokeRole
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleChangeRequest' }
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{userId}/enabled:
    post:
      tags: [Users]
      summary: Enable or disable a user
      description: A reason is required. The last enabled administrator cannot be disabled (`409`).
      operationId: setUserEnabled
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnabledRequest' }
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      operationId: listIncidents
      x-status: implemented
      x-roles: [OPERATIONS_SUPERVISOR, OPERATOR, MAINTENANCE_TECHNICIAN]
      parameters:
        - name: status
          in: query
          required: false
          description: Repeat to filter by several statuses.
          schema:
            type: array
            items: { $ref: '#/components/schemas/IncidentStatus' }
        - name: priority
          in: query
          required: false
          description: Repeat to filter by several priorities.
          schema:
            type: array
            items: { $ref: '#/components/schemas/Priority' }
        - name: assetId
          in: query
          required: false
          schema: { type: string }
        - name: sensorId
          in: query
          required: false
          schema: { type: string }
        - name: createdFrom
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          required: false
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of incidents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Incidents]
      summary: Create an incident (technical entry point)
      description: |
        A technical entry point that exists only when the Gateway runs with
        `COLDGUARD_TECHNICAL_ENDPOINTS_ENABLED=true` (on in local Compose). Real incidents are created
        from telemetry. An equivalent incident that is still open answers `409 INCIDENT_ALREADY_EXISTS`
        with the `existingIncidentId`.
      operationId: createIncident
      x-status: implemented
      x-roles: [PLATFORM_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIncidentRequest' }
      responses:
        '201':
          description: Created
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /incidents/{incidentId}:
    get:
      tags: [Incidents]
      summary: Get an incident
      operationId: getIncident
      x-status: implemented
      x-roles: [OPERATIONS_SUPERVISOR, OPERATOR, MAINTENANCE_TECHNICIAN]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: The incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /incidents/{incidentId}/acknowledgement:
    post:
      tags: [Incidents]
      summary: Acknowledge an incident
      description: Moves `CREATED` to `ACKNOWLEDGED`. A second acknowledgement answers `409 INCIDENT_ALREADY_ACKNOWLEDGED`.
      operationId: acknowledgeIncident
      x-status: implemented
      x-roles: [OPERATIONS_SUPERVISOR]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcknowledgementRequest' }
      responses:
        '200':
          description: The updated incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /incidents/{incidentId}/escalation:
    post:
      tags: [Incidents]
      summary: Escalate an incident
      description: A reason is required.
      operationId: escalateIncident
      x-status: implemented
      x-roles: [OPERATIONS_SUPERVISOR]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EscalationRequest' }
      responses:
        '200':
          description: The updated incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /incidents/{incidentId}/close:
    post:
      tags: [Incidents]
      summary: Close an incident
      description: |
        Only the maintenance technician closes an incident. The cause and the resolution comment
        are required and are kept with the incident (`cause`, `resolutionComment`, `closedAt`, `closedBy`).
      operationId: closeIncident
      x-status: implemented
      x-roles: [MAINTENANCE_TECHNICIAN]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CloseIncidentRequest' }
      responses:
        '200':
          description: The closed incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The `accessToken` returned by `POST /auth/login`.
...
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page number.
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      description: Page size; 0 or omitted selects the server default (20); the maximum is 100.
      schema: { type: integer, minimum: 0, maximum: 100, default: 0 }
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: |
        The request is not valid. `code` is `INVALID_REQUEST` (with `errors` naming each invalid field,
        never its value), `MALFORMED_REQUEST_BODY`, or a business code of the service.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthorized:
      description: No valid token, or `INVALID_CREDENTIALS` on login. Sign in again.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: The token is valid but its roles do not allow this operation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: |
        The resource does not exist (an id that is not valid cannot exist either). `code` names it:
        `ASSET_NOT_FOUND`, `SITE_NOT_FOUND`, `SENSOR_NOT_FOUND`, `ORGANIZATION_NOT_FOUND`,
        `PROFILE_NOT_FOUND`, `USER_NOT_FOUND`, `INCIDENT_NOT_FOUND`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: |
        The state does not allow it, or the data clashes. `CONCURRENT_MODIFICATION` means someone else
        changed the resource: reload it and retry. Others: `SENSOR_SERIAL_DUPLICATED`,
        `ORGANIZATION_NAME_DUPLICATED`, `SITE_NAME_DUPLICATED`, `USER_ALREADY_EXISTS`,
        `USER_STATE_CONFLICT`, `SENSOR_TRANSITION_NOT_ALLOWED`, `CALIBRATION_EVIDENCE_REQUIRED`,
        `CALIBRATION_EXPIRED`, `CALIBRATION_VALIDITY_NOT_CONFIGURED`, `REASSIGNMENT_NOT_ALLOWED`,
        `INCIDENT_ALREADY_EXISTS`, `INCIDENT_ALREADY_ACKNOWLEDGED`, `INCIDENT_ALREADY_CLOSED`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
    ProblemDetail:
      type: object
      required: [title, status, code]
      properties:
        type: { type: string, example: "about:blank" }
        title: { type: string, example: Conflict }
        status: { type: integer, example: 409 }
        detail: { type: string }
        instance: { type: string, description: Path of the request. }
        code:
          type: string
          description: Stable machine-readable code; branch on this, not on `detail`.
          example: SENSOR_TRANSITION_NOT_ALLOWED
        correlationId: { type: string }
        existingIncidentId: { type: string, description: "Only on `INCIDENT_ALREADY_EXISTS`: the incident that is already open." }
        errors:
          type: array
          description: Present on `INVALID_REQUEST`; one entry per invalid field.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, example: reason }
              message: { type: string, example: must not be blank }
    PageInfo:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
    Role:
      type: string
      enum: [OPERATIONS_SUPERVISOR, OPERATOR, MAINTENANCE_TECHNICIAN, AUDITOR, PLATFORM_ADMIN]
...
    User:
      type: object
      required: [userId, username, email, displayName, roles, enabled, createdAt]
      properties:
        userId: { type: string }
        username: { type: string }
        email: { type: string }
        displayName: { type: string }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/User' }
        page: { $ref: '#/components/schemas/PageInfo' }
    CreateUserRequest:
      type: object
      required: [username, email, displayName, initialPassword, roles]
      properties:
        username: { type: string }
        email: { type: string }
        displayName: { type: string }
        initialPassword: { type: string, format: password }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
    RoleChangeRequest:
      type: object
      required: [role, reason]
      properties:
        role: { $ref: '#/components/schemas/Role' }
        reason: { type: string }
    EnabledRequest:
      type: object
      required: [enabled, reason]
      properties:
        enabled: { type: boolean }
        reason: { type: string }
...
    IncidentStatus:
      type: string
      enum: [CREATED, ACKNOWLEDGED, ESCALATED, CLOSED]
    Priority:
      type: string
      enum: [P1, P2, P3, P4]
    Impact:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    Urgency:
      type: string
      enum: [LOW, MEDIUM, HIGH, IMMEDIATE]
    Incident:
      type: object
      required: [id, status, assetId, sensorId, anomalyType, impact, urgency, priority, createdAt, escalationCount, occurrenceCount]
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        assetId: { type: string }
        sensorId: { type: string }
        anomalyType: { type: string, example: HIGH_TEMPERATURE }
        impact: { $ref: '#/components/schemas/Impact' }
        urgency: { $ref: '#/components/schemas/Urgency' }
        priority: { $ref: '#/components/schemas/Priority' }
        createdAt: { type: string, format: date-time }
        ackDueAt: { type: string, format: date-time, nullable: true }
        resolveDueAt: { type: string, format: date-time, nullable: true, description: Absent when the priority has no resolution target. }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        acknowledgedBy: { type: string, nullable: true }
        lastEscalatedAt: { type: string, format: date-time, nullable: true }
        escalationCount: { type: integer }
        closedAt: { type: string, format: date-time, nullable: true }
        closedBy: { type: string, nullable: true }
        cause: { type: string, nullable: true }
        resolutionComment: { type: string, nullable: true }
        occurrenceCount: { type: integer, description: Equivalent anomalies seen while the incident is open. }
        lastOccurrenceAt: { type: string, format: date-time, nullable: true }
    IncidentPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Incident' }
        page: { $ref: '#/components/schemas/PageInfo' }
    AcknowledgementRequest:
      type: object
      properties:
        note: { type: string, maxLength: 500 }
    EscalationRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 500 }
    CloseIncidentRequest:
      type: object
      required: [cause, resolutionComment]
      properties:
        cause: { type: string, maxLength: 500 }
        resolutionComment: { type: string, maxLength: 2000 }
    CreateIncidentRequest:
      type: object
      required: [assetId, assetCriticality, sensorId, anomalyType, magnitude]
      properties:
        assetId: { type: string, maxLength: 100 }
        assetCriticality: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        sensorId: { type: string, maxLength: 100 }
        anomalyType: { type: string, maxLength: 100, example: TEMPERATURE_ABOVE_MAX }
        magnitude: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        persistent: { type: boolean, default: false }
...